Extend integration tests

+++ transmart-core-db/grails-app/domain/org/transmartproject/db/i2b2data/VisitDimension.groovy
@@ -5,7 +5,7 @@ import groovy.transform.EqualsAndHashCode
 @EqualsAndHashCode(includes = [ 'encounterNum', 'patient'])
 class VisitDimension implements Serializable{
 
-    Long                encounterNum
+    BigDecimal          encounterNum
     PatientDimension    patient
     String              activeStatusCd
     Date                startDate
